change data type from int to double

+++ ITI_Interview/Program.cs
@@ -3,9 +3,9 @@ internal class Program
 {
     private static void Main(string[] args)
     {
-        int FirstNumber = 10;
-        int SecondNumber = 20;
-        int Sum = FirstNumber + SecondNumber;
+        double FirstNumber = 10;
+        double SecondNumber = 20;
+        double Sum = FirstNumber + SecondNumber;
         Console.WriteLine("The sum of {0} and {1} is {2}", FirstNumber, SecondNumber, Sum);
